+++ spec/models/team_spec.rb
@@ -233,9 +233,11 @@ describe Team do
   context 'team' do
     let(:tz) { 'Eastern Time (US & Canada)' }
     let(:t_in_time_zone) { Time.now.utc.in_time_zone(tz) }
+    let(:eight_am) { 8 * 60 * 60 }
     let(:wday) { t_in_time_zone.wday }
     let(:beginning_of_day) { t_in_time_zone.beginning_of_day }
-    let(:team) { Fabricate(:team, sup_wday: wday, sup_time_of_day: 0, sup_tz: tz) }
+    let(:team) { Fabricate(:team, sup_wday: wday, sup_time_of_day: eight_am, sup_tz: tz) }
+    let(:on_time_sup) { beginning_of_day + team.sup_time_of_day }
 
     describe '#sync!' do
       let(:member_default_attr) do
@@ -353,7 +355,7 @@ describe Team do
 
     describe '#sup?' do
       before do
-        Timecop.travel(beginning_of_day + (8 * 60 * 60))
+        Timecop.travel(on_time_sup)
         allow(team).to receive(:sync!)
       end
 
@@ -363,7 +365,7 @@ describe Team do
         end
       end
 
-      context 'with a round' do
+      context 'with a round on time' do
         before do
           team.sup!
         end
@@ -374,7 +376,7 @@ describe Team do
 
         context 'after less than a week' do
           before do
-            Timecop.travel(Time.now.utc + 6.days)
+            Timecop.travel(on_time_sup + 6.days)
           end
 
           it 'is false' do
@@ -384,7 +386,7 @@ describe Team do
 
         context 'after more than a week' do
           before do
-            Timecop.travel(Time.now.utc + 7.days)
+            Timecop.travel(on_time_sup + 7.days)
           end
 
           it 'is true' do
@@ -409,7 +411,7 @@ describe Team do
 
           context 'after more than a week' do
             before do
-              Timecop.travel(Time.now.utc + 7.days)
+              Timecop.travel(on_time_sup + 7.days)
             end
 
             it 'is true' do
@@ -419,7 +421,7 @@ describe Team do
 
           context 'after more than two weeks' do
             before do
-              Timecop.travel(Time.now.utc + 14.days)
+              Timecop.travel(on_time_sup + 14.days)
             end
 
             it 'is true' do
@@ -430,7 +432,101 @@ describe Team do
 
         context 'after more than a week on the wrong day of the week' do
           before do
-            Timecop.travel(Time.now.utc + 8.days)
+            Timecop.travel(on_time_sup + 8.days)
+          end
+
+          it 'is false' do
+            expect(team.sup?).to be false
+          end
+        end
+      end
+
+      context 'with a round delayed by an hour' do
+        before do
+          Timecop.freeze(on_time_sup + 1.hour) do
+            team.sup!
+          end
+        end
+
+        context 'after less than a week' do
+          before do
+            Timecop.travel(on_time_sup + 6.days)
+          end
+
+          it 'is false' do
+            expect(team.sup?).to be false
+          end
+        end
+
+        context 'before sup time a week later' do
+          before do
+            Timecop.travel(on_time_sup + 7.days - 1.hour)
+          end
+
+          it 'is false' do
+            expect(team.sup?).to be false
+          end
+        end
+
+        context 'on time a week later' do
+          before do
+            Timecop.travel(on_time_sup + 7.days)
+          end
+
+          it 'is true' do
+            expect(team.sup?).to be true
+          end
+        end
+
+        context 'after more than a week' do
+          before do
+            Timecop.travel(on_time_sup + 7.days + 1.hour)
+          end
+
+          it 'is true' do
+            expect(team.sup?).to be true
+          end
+
+          context 'and another round' do
+            before do
+              team.sup!
+            end
+
+            it 'is false' do
+              expect(team.sup?).to be false
+            end
+          end
+        end
+
+        context 'with a custom sup_every_n_weeks' do
+          before do
+            team.update_attributes!(sup_every_n_weeks: 2)
+          end
+
+          context 'after more than a week' do
+            before do
+              Timecop.travel(on_time_sup + 7.days)
+            end
+
+            it 'is true' do
+              expect(team.sup?).to be false
+            end
+          end
+
+          context 'after more than two weeks' do
+            before do
+              Timecop.travel(on_time_sup + 14.days)
+            end
+
+            it 'is true' do
+              expect(team.sup?).to be true
+            end
+          end
+        end
+
+        context 'after more than a week on the wrong day of the week' do
+          before do
+            Timecop.travel(on_time_sup + 8.days)
           end
 
           it 'is false' do